+++ config/common.go
@@ -5,6 +5,7 @@ import "fmt"
 type Common struct {
 	FieldSettings                      []string
 	WrapErrors                         bool
+	WrapErrorsUsing                    string
 	IgnoreUnexported                   bool
 	MatchIgnoreCase                    bool
 	IgnoreMissing                      bool
@@ -16,7 +17,15 @@ type Common struct {
 func parseCommon(c *Common, cmd, rest string) (fieldSetting bool, err error) {
 	switch cmd {
 	case "wrapErrors":
+		if c.WrapErrorsUsing != "" {
+			return false, fmt.Errorf("cannot be used in combination with wrapErrorsUsing")
+		}
 		c.WrapErrors, err = parseBool(rest)
+	case "wrapErrorsUsing":
+		if c.WrapErrors {
+			return false, fmt.Errorf("cannot be used in combination with wrapErrors")
+		}
+		c.WrapErrorsUsing, err = parseString(rest)
 	case "ignoreUnexported":
 		fieldSetting = true
 		c.IgnoreUnexported, err = parseBool(rest)