+++ docs/changelog.md
@@ -7,6 +7,8 @@ import GH from './GH.vue';
 ## Unreleased
 
 - Require go1.18 for building Goverter
+- Add current working directory `-cwd` option to [CLI](./reference/cli.md)
+- Add [`wrapErrorsUsing`](./reference/wrapErrorsUsing.md)
 
 ## v1.3.2